add inbox domain
