+++ src/domains/inbox/models/notion_inbox.ts
@@ -1,4 +1,4 @@
-import { NotionRow } from "../core/entities/notion_row";
+import { NotionRow } from "../../../core/entities/notion_row";
 
 export default interface NotionInbox extends NotionRow {
 	title: string;
