+++ src/domains/inbox/models/ticktick_task.ts
@@ -1,4 +1,4 @@
-import Task from "../core/entities/task";
+import Task from "../../../core/entities/task";
 
 export default interface TicktickTask extends Task {
 	source: "ticktick";
