+++ src/repositories/contracts/iNotion.ts
@@ -1,13 +1,8 @@
+import TicktickTask from "../../domains/inbox/models/ticktick_task";
+import NotionJournal from "../../domains/journal/models/notion_journal";
 import NotionBlog, { NotionBlogContent } from "../../entities/notion_blog";
-import NotionInbox from "../../entities/notion_inbox";
-import NotionJournal from "../../entities/notion_journal";
-import TicktickTask from "../../entities/ticktick_task";
 
 export default interface INotion {
-  getInbox(id: string): Promise<NotionInbox | undefined>;
-  addJournal(ticktickTasks: TicktickTask[]): Promise<NotionJournal>;
-  addScreenshotToInbox(url: string): Promise<NotionInbox>;
-  addToInbox(title: string, body?: string): Promise<NotionInbox>;
   listBlog(): Promise<NotionBlog[]>;
   getBlogContent(blog: NotionBlog): Promise<NotionBlogContent>;
