+++ src/repositories/firestore.ts
@@ -6,12 +6,8 @@ import {
   anyFirestoreToDate,
 } from "../core/utils";
 import pocket_article from "../entities/pocket_article";
-import syncedInboxes from "../entities/syncedInboxes";
 import IFirestore from "./contracts/iFirestore";
 
-const CONFIG = "/general/config";
-const TASKS = "/tasks";
-const SYNCEDINBOXES = "/synced_inboxes";
 const BLOG = "/blog/last_update";
 const MODE = "/mode/";
 
@@ -97,54 +93,5 @@ export default class Firestore implements IFirestore {
     }
   }
 
-  async addSyncedInboxes(syncedInboxes: syncedInboxes): Promise<void> {
-    await this.client.collection(SYNCEDINBOXES).add(syncedInboxes);
-  }
 
-
-  async getCompletedTasks(after?: Date) {
-    let ref = this.client.collection(TASKS).where("done", "==", true);
-    if (after) {
-      ref = ref.where("created", ">=", firestore.Timestamp.fromDate(after));
-    }
-
-    const query = await ref.get();
-
-    return query.docs.map((d) => d.data()) as (task & { done: true })[];
-  }
-
-
-
-  getTasks(): Promise<task[]> {
-    throw new Error("Method not implemented.");
-  }
-
-  async addTask(task: task): Promise<void> {
-    const exists = await this.getTask(task.id);
-
-    // FIXME is it goo to throw time here without information the rest of the system
-    const taskWithTime = {
-      ...task,
-      created: firestore.Timestamp.now(),
-    };
-
-    if (exists) {
-      await this.client.doc(`${TASKS}/${exists.id}`).update(taskWithTime);
-    } else {
-      await this.client.collection(TASKS).add(taskWithTime);
-    }
-  }
-
-  /**
-   *
-   * @param taskId it's not the document id, but rather a stored id withing the document it self
-   */
-  private async getTask(taskId: string) {
-    const ref = this.client.collection(TASKS).where("id", "==", taskId);
-    const query = await ref.get();
-
-    if (query.size) {
-      return query.docs[0];
-    }
-  }
 }
